backend/pyproject.toml
[project]
name = "find-assist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "fastapi>=0.135.1",
    "jsonpatch>=1.33",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-huggingface>=1.2.0",
    "langchain-pinecone>=0.2.13",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=0.2.0",
    "langsmith>=0.7.6",
    "openai-whisper>=20250625",
    "psycopg2-binary>=2.9.11",
    "pypdf>=6.7.2",
    "rich>=14.3.3",
    "sentence-transformers>=5.2.3",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.41.0",
    "python-multipart>=0.0.24",
]
